student profile editable, some clickthrough
