+++ student-info-system/schooladmin/templatetags/user_demographics.py
@@ -2,6 +2,7 @@ from django import template
 
 register = template.Library()
 
+
 @register.inclusion_tag('demo.html')
 def demographic_details(profile):
     return {
