+++ student-info-system/sis/tables/__init__.py
@@ -47,4 +47,3 @@ class NameColumn(ClassyColumn):
 
 class AbilityColumn(tables.BooleanColumn):
     header = "Enabled?"
-
