+++ student-info-system/student/templates/student/student_edit.html
@@ -1,0 +1,35 @@
+{% extends 'sis/subpage_base.html' %}
+{% load static %}
+{% load bootstrap3 %}
+
+{% block content %}
+  <h3>Edit "{{ user.username }}" / {{ user.get_full_name }}</h3>
+    <div class="editform">
+  <form method="post">
+    {% csrf_token %}
+      <table>{{ user_form.as_table }}</table>
+      <table>{{ profile_form.as_table }}</table>
+      <table id="demo_table">{{ demo_form.as_table }}</table>
+    <input
+        action="action"
+        onclick="window.history.go(-1); return false;"
+        type="submit"
+        value="Cancel" class="btn btn-info"
+    />
+    <input
+        type="reset" id="reset_button"
+        value="Reset" class="btn btn-info"
+    />
+    <input id="submitme" name="submitme"
+        action="action"
+        value="Save Changes" class="btn btn-primary submitme"
+    />
+  </form>
+    </div>
+<script>
+document.getElementById("submitme").addEventListener("click", function () {
+        this.form.submit();
+        return true;
+    });
+</script>
+{% endblock %}
